+++ tests/test_read.py
@@ -1,6 +1,7 @@
 import logging
 import re
 from pathlib import Path
+import sys
 
 import pytest
 from xarray import Dataset
@@ -51,19 +52,3 @@ def test_read_file_parser_inputs(file_path, parser):
     """Test if read.file can accept parsers as None, string and parser it self"""
     dataset = read.file(file_path, parser=parser)
     assert isinstance(dataset, Dataset), "Output isn't an xarray dataset"
-
-@pytest.mark.parametrize(
-    "file_path,parser",
-    [("tests/parsers_test_files/seabird/btl/MI18MHDR.btl", "seabird.btl")]
-)
-def test_read_file_unique_import(file_path,parser,caplog):
-    """Test that read.file only import the parser once"""
-    parser_module = parser.rsplit(".")[0]
-    with caplog.at_level(logging.DEBUG):
-        read.file(file_path, parser=parser)
-        assert f"Import module: ocean_data_parser.parsers.{parser_module}" in caplog.text
-        caplog.clear()
-        read.file(file_path, parser=parser)
-        assert f"Import module: ocean_data_parser.parsers.{parser_module}" not in caplog.text
-        assert f"Module already imported: ocean_data_parser.parsers.{parser_module}" in caplog.text
-